Add function to generate an upload target
This is currently pretty simplistic. For example, all parameters like the port
or baudrate are global and cannot be configured per target. This should
probably be fixed in the near future.

The function will give a warning and not create the target when avrdude is not
found or the port has not been set by the user. The reason that we only err on
a missing port and not on missing baudrate or programmer type is that assuming
a default port seems to be the most dangerous because we might accidentally
start talking to a random device. If the user has already consciously set the
port, we assume they have thought about whether or not to use the default
values for the other parameters.